Ask the packager for architecture on Debian platforms
Debain has some strage ideas about what the package platform
should be, and it doesn't line up well with the kernel machine
in many cases. Instead of guessing, or having a lot of logic to
fix the architecture, we'll just ask dpkg for the correct one.

This works great when you have something like an ARMv7 host
and you are compiling for the armhf and armel architectures.

Haven't figure out how to effectively duplicate this with RPM
since this uses fpm to generate the package and not all RPM
tools are guaranteed to be installed on the host.